D4: extract rps_tournament_cog views to views/rps/ package
Reduces cogs/rps_tournament_cog.py from 1308 → 918 LOC (~30% reduction)
via option (a): pull every discord.ui.View class out of the cog into
the dedicated views package. Tournament state machine, registration
logic, and prefix commands stay in the cog — they're tightly coupled
to ``self`` state and the lifecycle, where extraction would carry
real regression risk.

New layout:
  disbot/views/rps/
  ├── __init__.py        re-exports (constants + 5 view classes)
  ├── _helpers.py        _RPS_WINS, _RPS_EMOJI, _FREE_WIN constants +
  │                      _rps_pvp_pending shared registry
  ├── solo_play.py       _RpsView (single-player game)
  ├── registration.py    _RpsRegistrationView (tournament join button)
  ├── pvp_challenge.py   _RpsPvpChallengeView (accept/decline)
  ├── pvp_play.py        _RpsPvpPlayView (both players pick + resolve)
  └── move_picker.py     _RpsMovePickerView (ephemeral per-player picker)

The PvP flow's transitive dependencies (challenge → play_view →
move_picker → play_view.record_choice) use local imports inside button
callbacks to break the would-be circular module graph.

Back-compat re-exports: cogs.rps_tournament_cog re-imports every
constant and view class from views.rps, so external imports like
``from cogs.rps_tournament_cog import _RpsPvpPlayView`` continue to
resolve.

Persistence-rule check: rps_tournament_cog never registered a
PersistentView, never used the interaction_router, and the views all
use timeout-based ephemeral state — no SUBSYSTEM string, no
custom_id, no panel_anchors row references.

Verification: ruff/black/isort/mypy clean; pytest 436/436 green.
