+++ disbot/cogs/rps_tournament_cog.py
@@ -12,11 +12,24 @@ from services import economy_service
 from utils import db as global_db
 from utils.channels import cleanup_category, create_private_channel
 from utils.settings_keys import ACTIVE_TOURNAMENT
-from utils.ui_constants import ERROR_COLOR, GAME_COLOR, INFO_COLOR, SUCCESS_COLOR
+from utils.ui_constants import GAME_COLOR, INFO_COLOR
+
+# Views + shared constants moved to views/rps/ during D4 — re-exported
+# below for backward compatibility with any external import of these
+# private names.
+from views.rps import (  # noqa: F401 — re-exported for back-compat
+    _FREE_WIN,
+    _RPS_EMOJI,
+    _RPS_WINS,
+    _rps_pvp_pending,
+    _RpsMovePickerView,
+    _RpsPvpChallengeView,
+    _RpsPvpPlayView,
+    _RpsRegistrationView,
+    _RpsView,
+)
 
 logger = logging.getLogger("bot")
-
-_FREE_WIN = 30  # coins for free-play win
 
 
 class RPSTournamentCog(commands.Cog, name="Rock-Paper-Scissors Tournament"):  # type: ignore[call-arg]
@@ -899,409 +912,6 @@ class RPSTournamentCog(commands.Cog, name="Rock-Paper-Scissors Tournament"):  # 
         )
         msg = await ctx.send(embed=embed, view=view)
         view.message = msg
-
-
-# ---------------------------------------------------------------------------
-# Quick-play RPS View
-# ---------------------------------------------------------------------------
-
-_RPS_WINS = {"rock": "scissors", "scissors": "paper", "paper": "rock"}
-_RPS_EMOJI = {"rock": "🪨", "paper": "📄", "scissors": "✂️"}
-
-
-class _RpsView(discord.ui.View):
-    def __init__(self, user: discord.Member, guild_id: int, bet: int):
-        super().__init__(timeout=60)
-        self.user = user
-        self.guild_id = guild_id
-        self.bet = bet
-        self.message: discord.Message | None = None
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if interaction.user.id != self.user.id:
-            await interaction.response.send_message(
-                "This game isn't yours.",
-                ephemeral=True,
-            )
-            return False
-        return True
-
-    async def _play(self, interaction: discord.Interaction, player_move: str):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-
-        bot_move = random.choice(["rock", "paper", "scissors"])
-        pe, be = _RPS_EMOJI[player_move], _RPS_EMOJI[bot_move]
-
-        if player_move == bot_move:
-            result = "🤝 Tie!"
-            coin_delta = 0
-            color = GAME_COLOR
-        elif _RPS_WINS[player_move] == bot_move:
-            payout = self.bet if self.bet else _FREE_WIN
-            result = f"🎉 You win! +{payout} 🪙"
-            coin_delta = payout
-            color = SUCCESS_COLOR
-        else:
-            loss = -self.bet if self.bet else 0
-            result = f"😞 Bot wins. {loss} 🪙" if self.bet else "😞 Bot wins."
-            coin_delta = loss
-            color = ERROR_COLOR
-
-        # Single-player RPS: bet was not pre-deducted, outcome is applied
-        # directly. Positive coin_delta = win (credit), negative = loss (debit
-        # with overdraft to preserve the original floor-at-zero behaviour).
-        if coin_delta > 0:
-            new_bal = await economy_service.credit(
-                self.guild_id,
-                self.user.id,
-                coin_delta,
-                reason="rps:solo_win",
-                actor_id=self.user.id,
-            )
-        elif coin_delta < 0:
-            new_bal = await economy_service.debit(
-                self.guild_id,
-                self.user.id,
-                -coin_delta,
-                reason="rps:solo_loss",
-                actor_id=self.user.id,
-                allow_overdraft=True,
-            )
-        else:
-            new_bal = await global_db.get_coins(self.user.id, self.guild_id)
-        embed = discord.Embed(
-            title="✂️ Rock · Paper · Scissors",
-            description=(
-                f"You: **{player_move}** {pe}  vs  Bot: **{bot_move}** {be}\n\n"
-                f"{result}\n"
-                f"Balance: **{new_bal}** 🪙"
-            ),
-            color=color,
-        )
-        await interaction.response.edit_message(embed=embed, view=self)
-        self.stop()
-
-    @discord.ui.button(label="Rock", emoji="🪨", style=discord.ButtonStyle.grey)
-    async def rock(self, i: discord.Interaction, _: discord.ui.Button):
-        await self._play(i, "rock")
-
-    @discord.ui.button(label="Paper", emoji="📄", style=discord.ButtonStyle.grey)
-    async def paper(self, i: discord.Interaction, _: discord.ui.Button):
-        await self._play(i, "paper")
-
-    @discord.ui.button(label="Scissors", emoji="✂️", style=discord.ButtonStyle.grey)
-    async def scissors(self, i: discord.Interaction, _: discord.ui.Button):
-        await self._play(i, "scissors")
-
-    async def on_timeout(self):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        try:
-            await self.message.edit(content="Game timed out.", view=self)
-        except Exception:
-            pass
-
-
-# ---------------------------------------------------------------------------
-# RPS Registration View (button-based join)
-# ---------------------------------------------------------------------------
-
-
-class _RpsRegistrationView(discord.ui.View):
-    def __init__(self, cog: RPSTournamentCog):
-        super().__init__(timeout=None)  # lives until tournament starts
-        self.cog = cog
-
-    @discord.ui.button(
-        label="Join Tournament",
-        style=discord.ButtonStyle.green,
-        emoji="✅",
-    )
-    async def join_btn(self, interaction: discord.Interaction, _: discord.ui.Button):
-        cog = self.cog
-        if not cog.registration_active:
-            await interaction.response.send_message(
-                "Registration is no longer open.",
-                ephemeral=True,
-            )
-            return
-        guild_id = interaction.guild_id or 0
-        ok = await cog.try_register_player(interaction.user, guild_id)
-        if ok:
-            await interaction.response.send_message(
-                f"✅ Registered! ({len(cog.players)} player(s) so far)",
-                ephemeral=True,
-            )
-        else:
-            bal = await global_db.get_coins(interaction.user.id, guild_id)
-            if cog.entry_fee > 0 and bal < cog.entry_fee:
-                await interaction.response.send_message(
-                    f"❌ Need **{cog.entry_fee}** 🪙 to enter (you have {bal}).",
-                    ephemeral=True,
-                )
-            else:
-                await interaction.response.send_message(
-                    "You're already registered!",
-                    ephemeral=True,
-                )
-
-
-# ---------------------------------------------------------------------------
-# RPS PvP Challenge
-# ---------------------------------------------------------------------------
-
-_rps_pvp_pending: dict[frozenset, dict] = (
-    {}
-)  # {p1,p2} → {choices, guild_id, bet, channel_id}
-
-
-class _RpsPvpChallengeView(discord.ui.View):
-    def __init__(
-        self,
-        challenger: discord.Member,
-        opponent: discord.Member,
-        guild_id: int,
-        bet: int,
-    ):
-        super().__init__(timeout=60)
-        self.challenger = challenger
-        self.opponent = opponent
-        self.guild_id = guild_id
-        self.bet = bet
-        self.message: discord.Message | None = None
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if interaction.user.id != self.opponent.id:
-            await interaction.response.send_message(
-                "This challenge isn't for you.",
-                ephemeral=True,
-            )
-            return False
-        return True
-
-    @discord.ui.button(label="Accept", style=discord.ButtonStyle.green, emoji="✅")
-    async def accept(self, interaction: discord.Interaction, _: discord.ui.Button):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        await interaction.response.edit_message(
-            content="✅ Challenge accepted — both players, choose your move!",
-            view=self,
-        )
-        key = frozenset({self.challenger.id, self.opponent.id})
-        _rps_pvp_pending[key] = {
-            "choices": {},
-            "guild_id": self.guild_id,
-            "bet": self.bet,
-            "channel_id": interaction.channel_id,
-        }
-        # Send ephemeral choose-views to both players
-        ch = interaction.channel
-        play_view = _RpsPvpPlayView(
-            self.challenger,
-            self.opponent,
-            self.guild_id,
-            self.bet,
-            ch,  # type: ignore[arg-type]
-        )
-        await ch.send(  # type: ignore[union-attr]
-            f"{self.challenger.mention} {self.opponent.mention} — click below to pick your move (only you can see your choice):",
-            view=play_view,
-        )
-        self.stop()
-
-    @discord.ui.button(label="Decline", style=discord.ButtonStyle.red, emoji="❌")
-    async def decline(self, interaction: discord.Interaction, _: discord.ui.Button):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        await interaction.response.edit_message(
-            content=f"❌ {self.opponent.display_name} declined the challenge.",
-            view=self,
-        )
-        self.stop()
-
-    async def on_timeout(self):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        try:
-            await self.message.edit(content="⏰ Challenge timed out.", view=self)
-        except Exception:
-            pass
-
-
-class _RpsPvpPlayView(discord.ui.View):
-    """Visible to the channel; each player clicks to get their ephemeral move picker."""
-
-    def __init__(
-        self,
-        p1: discord.Member,
-        p2: discord.Member,
-        guild_id: int,
-        bet: int,
-        channel: discord.TextChannel,
-    ):
-        super().__init__(timeout=60)
-        self.p1 = p1
-        self.p2 = p2
-        self.guild_id = guild_id
-        self.bet = bet
-        self.channel = channel
-        self.choices: dict[int, str] = {}
-        self.message: discord.Message | None = None
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if interaction.user.id not in (self.p1.id, self.p2.id):
-            await interaction.response.send_message(
-                "You're not part of this match.",
-                ephemeral=True,
-            )
-            return False
-        return True
-
-    @discord.ui.button(
-        label="Pick your move",
-        style=discord.ButtonStyle.blurple,
-        emoji="✂️",
-    )
-    async def pick(self, interaction: discord.Interaction, _: discord.ui.Button):
-        if interaction.user.id in self.choices:
-            await interaction.response.send_message(
-                "You already picked!",
-                ephemeral=True,
-            )
-            return
-        picker_view = _RpsMovePickerView(interaction.user.id, self)
-        await interaction.response.send_message(
-            "Choose your move — only you can see this:",
-            view=picker_view,
-            ephemeral=True,
-        )
-
-    async def record_choice(self, user_id: int, move: str):
-        self.choices[user_id] = move
-        if len(self.choices) == 2:
-            await self._resolve()
-
-    async def _resolve(self):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        try:
-            await self.message.edit(view=self)
-        except Exception:
-            pass
-        self.stop()
-
-        m1 = self.choices.get(self.p1.id, "forfeit")
-        m2 = self.choices.get(self.p2.id, "forfeit")
-
-        def _wins(a, b):
-            return {"rock": "scissors", "scissors": "paper", "paper": "rock"}.get(
-                a,
-            ) == b
-
-        e = {"rock": "🪨", "paper": "📄", "scissors": "✂️", "forfeit": "❌"}
-
-        if m1 == "forfeit" and m2 == "forfeit":
-            result, coin_delta, winner_id = "🤝 Both forfeited.", 0, None
-        elif m1 == "forfeit":
-            result, coin_delta, winner_id = (
-                f"{self.p2.mention} wins (opponent forfeited)!",
-                self.bet,
-                self.p2.id,
-            )
-        elif m2 == "forfeit":
-            result, coin_delta, winner_id = (
-                f"{self.p1.mention} wins (opponent forfeited)!",
-                self.bet,
-                self.p1.id,
-            )
-        elif m1 == m2:
-            result, coin_delta, winner_id = "🤝 Tie! No coins exchanged.", 0, None
-        elif _wins(m1, m2):
-            result, coin_delta, winner_id = (
-                f"🎉 {self.p1.mention} wins!",
-                self.bet,
-                self.p1.id,
-            )
-        else:
-            result, coin_delta, winner_id = (
-                f"🎉 {self.p2.mention} wins!",
-                self.bet,
-                self.p2.id,
-            )
-
-        if coin_delta and winner_id:
-            loser_id = self.p2.id if winner_id == self.p1.id else self.p1.id
-            payout = coin_delta if coin_delta else _FREE_WIN
-            # Preserves prior semantics: winner always credited the full
-            # payout; loser debited down to floor-zero if short (overdraft
-            # allowed). Using economy_service.transfer would atomically
-            # reject the move on insufficient loser balance — that is the
-            # safer pattern but a behaviour change; revisit when game
-            # rules formalise bet escrow.
-            await economy_service.credit(
-                self.guild_id,
-                winner_id,
-                payout,
-                reason="rps:pvp_win",
-            )
-            await economy_service.debit(
-                self.guild_id,
-                loser_id,
-                payout,
-                reason="rps:pvp_loss",
-                allow_overdraft=True,
-            )
-
-        embed = discord.Embed(
-            title="✂️ RPS PvP Result",
-            description=(
-                f"{self.p1.mention}: **{m1}** {e.get(m1, '')}\n"
-                f"{self.p2.mention}: **{m2}** {e.get(m2, '')}\n\n"
-                f"{result}"
-            ),
-            color=SUCCESS_COLOR if winner_id else GAME_COLOR,
-        )
-        await self.channel.send(embed=embed)
-
-    async def on_timeout(self):
-        # Anyone who didn't choose forfeits
-        for pid in (self.p1.id, self.p2.id):
-            if pid not in self.choices:
-                self.choices[pid] = "forfeit"
-        if len(self.choices) == 2:
-            await self._resolve()
-
-
-class _RpsMovePickerView(discord.ui.View):
-    """Ephemeral view for picking a move in PvP."""
-
-    def __init__(self, user_id: int, parent: _RpsPvpPlayView):
-        super().__init__(timeout=55)
-        self.user_id = user_id
-        self.parent = parent
-
-    @discord.ui.button(label="Rock", emoji="🪨", style=discord.ButtonStyle.grey)
-    async def rock(self, i: discord.Interaction, _):
-        await self._pick(i, "rock")
-
-    @discord.ui.button(label="Paper", emoji="📄", style=discord.ButtonStyle.grey)
-    async def paper(self, i: discord.Interaction, _):
-        await self._pick(i, "paper")
-
-    @discord.ui.button(label="Scissors", emoji="✂️", style=discord.ButtonStyle.grey)
-    async def scissors(self, i: discord.Interaction, _):
-        await self._pick(i, "scissors")
-
-    async def _pick(self, interaction: discord.Interaction, move: str):
-        for item in self.children:
-            item.disabled = True  # type: ignore[attr-defined]
-        await interaction.response.edit_message(
-            content=f"You chose **{move}** — waiting for opponent…",
-            view=self,
-        )
-        self.stop()
-        await self.parent.record_choice(self.user_id, move)
 
 
 async def setup(bot):
